+++ exporters/otlp/otlp.go
@@ -20,12 +20,9 @@ package otlp // import "go.opentelemetry.io/otel/exporters/otlp"
 import (
 	"context"
 	"errors"
-	"fmt"
 	"sync"
-	"unsafe"
 
 	"google.golang.org/grpc"
-	"google.golang.org/grpc/metadata"
 
 	colmetricpb "go.opentelemetry.io/otel/exporters/otlp/internal/opentelemetry-proto-gen/collector/metrics/v1"
 	coltracepb "go.opentelemetry.io/otel/exporters/otlp/internal/opentelemetry-proto-gen/collector/trace/v1"
@@ -43,22 +40,16 @@ type Exporter struct {
 	// mu protects the non-atomic and non-channel variables
 	mu sync.RWMutex
 	// senderMu protects the concurrent unsafe sends on the shared gRPC client connection.
-	senderMu          sync.Mutex
-	started           bool
-	traceExporter     coltracepb.TraceServiceClient
-	metricExporter    colmetricpb.MetricsServiceClient
-	grpcClientConn    *grpc.ClientConn
-	lastConnectErrPtr unsafe.Pointer
+	senderMu       sync.Mutex
+	started        bool
+	traceExporter  coltracepb.TraceServiceClient
+	metricExporter colmetricpb.MetricsServiceClient
+	cc             *grpcConnection
 
-	startOnce      sync.Once
-	stopOnce       sync.Once
-	stopCh         chan struct{}
-	disconnectedCh chan bool
+	startOnce sync.Once
+	stopOnce  sync.Once
 
-	backgroundConnectionDoneCh chan bool
-
-	c        config
-	metadata metadata.MD
+	exportKindSelector metricsdk.ExportKindSelector
 }
 
 var _ tracesdk.SpanExporter = (*Exporter)(nil)
@@ -82,9 +73,9 @@ func newConfig(opts ...ExporterOption) config {
 }
 
 // NewExporter constructs a new Exporter and starts it.
-func NewExporter(opts ...ExporterOption) (*Exporter, error) {
+func NewExporter(ctx context.Context, opts ...ExporterOption) (*Exporter, error) {
 	exp := NewUnstartedExporter(opts...)
-	if err := exp.Start(); err != nil {
+	if err := exp.Start(ctx); err != nil {
 		return nil, err
 	}
 	return exp, nil
@@ -93,19 +84,29 @@ func NewExporter(opts ...ExporterOption) (*Exporter, error) {
 // NewUnstartedExporter constructs a new Exporter and does not start it.
 func NewUnstartedExporter(opts ...ExporterOption) *Exporter {
 	e := new(Exporter)
-	e.c = newConfig(opts...)
-	if len(e.c.headers) > 0 {
-		e.metadata = metadata.New(e.c.headers)
-	}
+	cfg := newConfig(opts...)
+	e.exportKindSelector = cfg.exportKindSelector
+	e.cc = newGRPCConnection(cfg, e.handleNewConnection)
 	return e
 }
 
+func (e *Exporter) handleNewConnection(cc *grpc.ClientConn) error {
+	e.mu.Lock()
+	defer e.mu.Unlock()
+	if cc != nil {
+		e.metricExporter = colmetricpb.NewMetricsServiceClient(cc)
+		e.traceExporter = coltracepb.NewTraceServiceClient(cc)
+	} else {
+		e.metricExporter = nil
+		e.traceExporter = nil
+	}
+	return nil
+}
+
 var (
-	errAlreadyStarted  = errors.New("already started")
-	errNotStarted      = errors.New("not started")
-	errDisconnected    = errors.New("exporter disconnected")
-	errStopped         = errors.New("exporter stopped")
-	errContextCanceled = errors.New("context canceled")
+	errNoClient       = errors.New("no client")
+	errAlreadyStarted = errors.New("already started")
+	errDisconnected   = errors.New("exporter disconnected")
 )
 
 // Start dials to the collector, establishing a connection to it. It also
@@ -113,108 +114,25 @@ var (
 // messages that consist of the node identifier. Start invokes a background
 // connector that will reattempt connections to the collector periodically
 // if the connection dies.
-func (e *Exporter) Start() error {
+func (e *Exporter) Start(ctx context.Context) error {
 	var err = errAlreadyStarted
 	e.startOnce.Do(func() {
 		e.mu.Lock()
 		e.started = true
-		e.disconnectedCh = make(chan bool, 1)
-		e.stopCh = make(chan struct{})
-		e.backgroundConnectionDoneCh = make(chan bool)
 		e.mu.Unlock()
 
-		// An optimistic first connection attempt to ensure that
-		// applications under heavy load can immediately process
-		// data. See https://github.com/census-ecosystem/opencensus-go-exporter-ocagent/pull/63
-		if err := e.connect(); err == nil {
-			e.setStateConnected()
-		} else {
-			e.setStateDisconnected(err)
-		}
-		go e.indefiniteBackgroundConnection()
-
 		err = nil
+		e.cc.startConnection(ctx)
 	})
 
 	return err
-}
-
-func (e *Exporter) prepareCollectorAddress() string {
-	if e.c.collectorAddr != "" {
-		return e.c.collectorAddr
-	}
-	return fmt.Sprintf("%s:%d", DefaultCollectorHost, DefaultCollectorPort)
-}
-
-func (e *Exporter) enableConnections(cc *grpc.ClientConn) error {
-	e.mu.RLock()
-	started := e.started
-	e.mu.RUnlock()
-
-	if !started {
-		return errNotStarted
-	}
-
-	e.mu.Lock()
-	// If previous clientConn is same as the current then just return.
-	// This doesn't happen right now as this func is only called with new ClientConn.
-	// It is more about future-proofing.
-	if e.grpcClientConn == cc {
-		e.mu.Unlock()
-		return nil
-	}
-	// If the previous clientConn was non-nil, close it
-	if e.grpcClientConn != nil {
-		_ = e.grpcClientConn.Close()
-	}
-	e.grpcClientConn = cc
-	e.traceExporter = coltracepb.NewTraceServiceClient(cc)
-	e.metricExporter = colmetricpb.NewMetricsServiceClient(cc)
-	e.mu.Unlock()
-
-	return nil
-}
-
-func (e *Exporter) contextWithMetadata(ctx context.Context) context.Context {
-	if e.metadata.Len() > 0 {
-		return metadata.NewOutgoingContext(ctx, e.metadata)
-	}
-	return ctx
-}
-
-func (e *Exporter) dialToCollector() (*grpc.ClientConn, error) {
-	addr := e.prepareCollectorAddress()
-
-	dialOpts := []grpc.DialOption{}
-	if e.c.grpcServiceConfig != "" {
-		dialOpts = append(dialOpts, grpc.WithDefaultServiceConfig(e.c.grpcServiceConfig))
-	}
-	if e.c.clientCredentials != nil {
-		dialOpts = append(dialOpts, grpc.WithTransportCredentials(e.c.clientCredentials))
-	} else if e.c.canDialInsecure {
-		dialOpts = append(dialOpts, grpc.WithInsecure())
-	}
-	if e.c.compressor != "" {
-		dialOpts = append(dialOpts, grpc.WithDefaultCallOptions(grpc.UseCompressor(e.c.compressor)))
-	}
-	if len(e.c.grpcDialOptions) != 0 {
-		dialOpts = append(dialOpts, e.c.grpcDialOptions...)
-	}
-
-	ctx := e.contextWithMetadata(context.Background())
-	return grpc.DialContext(ctx, addr, dialOpts...)
-}
-
-// closeStopCh is used to wrap the exporters stopCh channel closing for testing.
-var closeStopCh = func(stopCh chan struct{}) {
-	close(stopCh)
 }
 
 // Shutdown closes all connections and releases resources currently being used
 // by the exporter. If the exporter is not started this does nothing.
 func (e *Exporter) Shutdown(ctx context.Context) error {
 	e.mu.RLock()
-	cc := e.grpcClientConn
+	cc := e.cc
 	started := e.started
 	e.mu.RUnlock()
 
@@ -225,23 +143,13 @@ func (e *Exporter) Shutdown(ctx context.Context) error {
 	var err error
 
 	e.stopOnce.Do(func() {
-		if cc != nil {
-			// Clean things up before checking this error.
-			err = cc.Close()
-		}
+		// Clean things up before checking this error.
+		err = cc.shutdown(ctx)
 
 		// At this point we can change the state variable started
 		e.mu.Lock()
 		e.started = false
 		e.mu.Unlock()
-		closeStopCh(e.stopCh)
-
-		// Ensure that the backgroundConnector returns
-		select {
-		case <-e.backgroundConnectionDoneCh:
-		case <-ctx.Done():
-			err = ctx.Err()
-		}
 	})
 
 	return err
@@ -251,16 +159,8 @@ func (e *Exporter) Shutdown(ctx context.Context) error {
 // interface. It transforms and batches metric Records into OTLP Metrics and
 // transmits them to the configured collector.
 func (e *Exporter) Export(parent context.Context, cps metricsdk.CheckpointSet) error {
-	// Unify the parent context Done signal with the exporter stopCh.
-	ctx, cancel := context.WithCancel(parent)
+	ctx, cancel := e.cc.contextWithStop(parent)
 	defer cancel()
-	go func(ctx context.Context, cancel context.CancelFunc) {
-		select {
-		case <-ctx.Done():
-		case <-e.stopCh:
-			cancel()
-		}
-	}(ctx, cancel)
 
 	// Hardcode the number of worker goroutines to 1. We later will
 	// need to see if there's a way to adjust that number for longer
@@ -270,32 +170,31 @@ func (e *Exporter) Export(parent context.Context, cps metricsdk.CheckpointSet) e
 		return err
 	}
 
-	if !e.connected() {
+	if !e.cc.connected() {
 		return errDisconnected
 	}
 
-	select {
-	case <-e.stopCh:
-		return errStopped
-	case <-ctx.Done():
-		return errContextCanceled
-	default:
+	err = func() error {
 		e.senderMu.Lock()
-		_, err := e.metricExporter.Export(e.contextWithMetadata(ctx), &colmetricpb.ExportMetricsServiceRequest{
+		defer e.senderMu.Unlock()
+		if e.metricExporter == nil {
+			return errNoClient
+		}
+		_, err := e.metricExporter.Export(e.cc.contextWithMetadata(ctx), &colmetricpb.ExportMetricsServiceRequest{
 			ResourceMetrics: rms,
 		})
-		e.senderMu.Unlock()
-		if err != nil {
-			return err
-		}
+		return err
+	}()
+	if err != nil {
+		e.cc.setStateDisconnected(err)
 	}
-	return nil
+	return err
 }
 
 // ExportKindFor reports back to the OpenTelemetry SDK sending this Exporter
 // metric telemetry that it needs to be provided in a cumulative format.
 func (e *Exporter) ExportKindFor(desc *metric.Descriptor, kind aggregation.Kind) metricsdk.ExportKind {
-	return e.c.exportKindSelector.ExportKindFor(desc, kind)
+	return e.exportKindSelector.ExportKindFor(desc, kind)
 }
 
 // ExportSpans exports a batch of SpanData.
@@ -304,28 +203,31 @@ func (e *Exporter) ExportSpans(ctx context.Context, sds []*tracesdk.SpanData) er
 }
 
 func (e *Exporter) uploadTraces(ctx context.Context, sdl []*tracesdk.SpanData) error {
-	select {
-	case <-e.stopCh:
+	ctx, cancel := e.cc.contextWithStop(ctx)
+	defer cancel()
+
+	if !e.cc.connected() {
 		return nil
-	default:
-		if !e.connected() {
-			return nil
-		}
+	}
 
-		protoSpans := transform.SpanData(sdl)
-		if len(protoSpans) == 0 {
-			return nil
-		}
+	protoSpans := transform.SpanData(sdl)
+	if len(protoSpans) == 0 {
+		return nil
+	}
 
+	err := func() error {
 		e.senderMu.Lock()
-		_, err := e.traceExporter.Export(e.contextWithMetadata(ctx), &coltracepb.ExportTraceServiceRequest{
+		defer e.senderMu.Unlock()
+		if e.traceExporter == nil {
+			return errNoClient
+		}
+		_, err := e.traceExporter.Export(e.cc.contextWithMetadata(ctx), &coltracepb.ExportTraceServiceRequest{
 			ResourceSpans: protoSpans,
 		})
-		e.senderMu.Unlock()
-		if err != nil {
-			e.setStateDisconnected(err)
-			return err
-		}
+		return err
+	}()
+	if err != nil {
+		e.cc.setStateDisconnected(err)
 	}
-	return nil
+	return err
 }